Update folder of client
