+++ client/src/App.jsx
@@ -1,0 +1,12 @@
+import {BrowserRouter, Routes, Route} from 'react-router-dom';
+import './App.css';
+
+function App() {
+  return (
+    <BrowserRouter>
+      <h1>hello there</h1>
+    </BrowserRouter>
+  );
+}
+
+export default App;
